daemon: Use systemd-provided sockets if available
At the moment, `crc daemon` must be run manually before one can use its
HTTP API, or when using usermode networking. Using systemd socket
activation would let systemd listen for HTTP/vsock connection attempts,
and tell it to automatically start the daemon when such a connection
attempt is made.

When using systemd socket activation
http://0pointer.de/blog/projects/socket-activation.html
the daemon should not create its own sockets for http/vsock
communication. Instead, systemd will provide already opened sockets when
starting the daemon, and the daemon is expected to reuse these.

This commit adds support for this by using go-systemd. vsock sockets
need to be special cased as they are not natively supported by the
golang net package, nor by go-systemd.

After this commit, running `crc daemon` on its own should still be
working as before.
Socket activation also requires some systemd unit files which
will be added in follow-up commits, so the daemon will only be
autostarted later on.

+++ cmd/crc/cmd/daemon_linux.go
@@ -1,16 +1,88 @@
 package cmd
 
 import (
+	"fmt"
 	"net"
 	"os"
 
 	"github.com/code-ready/crc/pkg/crc/constants"
 	"github.com/code-ready/crc/pkg/crc/logging"
 	"github.com/code-ready/gvisor-tap-vsock/pkg/transport"
+
+	"github.com/coreos/go-systemd/activation"
+	"github.com/mdlayher/vsock"
 )
 
+const (
+	vsockUnitName = "crc-vsock.socket"
+	httpUnitName  = "crc-http.socket"
+)
+
+var systemdListeners map[string][]net.Listener
+
+func init() {
+	// listenerWithNames() cannot be called multiple times
+	systemdListeners, _ = listenersWithNames()
+}
+
+// listenersWithNames maps a listener name to a set of net.Listener instances.
+// This is the same code as https://github.com/coreos/go-systemd/blob/main/activation/listeners.go
+// with support for vsock
+//
+// This function can only be called once, subsequent calls will always return an empty list
+func listenersWithNames() (map[string][]net.Listener, error) {
+	files := activation.Files(true)
+	listeners := map[string][]net.Listener{}
+
+	for _, f := range files {
+		pc, err := net.FileListener(f)
+		if err != nil {
+			logging.Infof("got error %t %v", err, err)
+			// net.FileListener does not support vsock, need to fallback to vsock-specific code
+			pc, err = vsock.FileListener(f)
+			if err != nil {
+				logging.Debugf("failed to create listener for %s: %v", f.Name(), err)
+				continue
+			}
+		}
+		current, ok := listeners[f.Name()]
+		if !ok {
+			listeners[f.Name()] = []net.Listener{pc}
+		} else {
+			listeners[f.Name()] = append(current, pc)
+		}
+		f.Close()
+	}
+	return listeners, nil
+}
+
+func getSystemdListener(unitName string) (net.Listener, error) {
+	listeners, found := systemdListeners[unitName]
+	if !found {
+		logging.Debugf("no listener for %s: %+v", unitName, systemdListeners)
+		// it's not an error if systemd provided no listener
+		return nil, nil
+	}
+	if len(listeners) != 1 {
+		logging.Debugf("unexpected number of sockets for %s (%d != 1)", unitName, len(listeners))
+		return nil, fmt.Errorf("unexpected number of sockets for %s (%d != 1)", unitName, len(listeners))
+	}
+
+	return listeners[0], nil
+}
+
 func vsockListener() (net.Listener, error) {
-	ln, err := transport.Listen(transport.DefaultURL)
+	ln, err := getSystemdListener(vsockUnitName)
+	if err != nil {
+		return nil, err
+	}
+	if ln != nil {
+		logging.Infof("using socket provided by %s", vsockUnitName)
+		return ln, nil
+	}
+
+	// no socket activation, we need to create the listener
+	ln, err = transport.Listen(transport.DefaultURL)
 	logging.Infof("listening %s", transport.DefaultURL)
 	if err != nil {
 		return nil, err
@@ -19,8 +91,19 @@ func vsockListener() (net.Listener, error) {
 }
 
 func httpListener() (net.Listener, error) {
+	// check for systemd socket-activation
+	ln, err := getSystemdListener(httpUnitName)
+	if err != nil {
+		return nil, err
+	}
+	if ln != nil {
+		logging.Infof("using socket provided by %s", httpUnitName)
+		return ln, nil
+	}
+
+	// no socket activation, we need to create the listener
 	_ = os.Remove(constants.DaemonHTTPSocketPath)
-	ln, err := net.Listen("unix", constants.DaemonHTTPSocketPath)
+	ln, err = net.Listen("unix", constants.DaemonHTTPSocketPath)
 	logging.Infof("listening %s", constants.DaemonHTTPSocketPath)
 	if err != nil {
 		return nil, err
